# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Task Management API
  description: RESTful API untuk sistem manajemen task dengan autentikasi JWT
  version: 1.0.0
  contact:
    name: API Support
    email: support@taskmanagement.com

servers:
  - url: http://localhost:3000/api
    description: Development server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token yang didapat dari endpoint /auth/login

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"

    Task:
      type: object
      properties:
        id:
          type: integer
          example: 1
        userId:
          type: integer
          example: 1
        title:
          type: string
          example: "Complete project documentation"
        description:
          type: string
          example: "Write comprehensive API documentation"
        priority:
          type: string
          enum: [low, medium, high]
          example: "high"
        completed:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Error:
      type: object
      properties:
        message:
          type: string
          example: "Error message"

paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Membuat akun user baru dengan validasi input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  minLength: 2
                  example: "John Doe"
                email:
                  type: string
                  format: email
                  example: "john@example.com"
                password:
                  type: string
                  minLength: 6
                  example: "password123"
      responses:
        '201':
          description: User berhasil didaftarkan
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User registered successfully"
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validasi gagal atau email sudah terdaftar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Autentikasi user dan mendapatkan JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "john@example.com"
                password:
                  type: string
                  example: "password123"
      responses:
        '200':
          description: Login berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Login successful"
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Email atau password salah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Email dan password wajib diisi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /tasks:
    get:
      tags:
        - Tasks
      summary: List all tasks
      description: Mendapatkan semua task milik user yang sedang login
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Daftar tasks berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Tasks retrieved successfully"
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  count:
                    type: integer
                    example: 5
        '401':
          description: Token tidak ada atau tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Tasks
      summary: Create new task
      description: Membuat task baru untuk user yang sedang login
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  minLength: 3
                  example: "Complete project documentation"
                description:
                  type: string
                  example: "Write comprehensive API documentation"
                priority:
                  type: string
                  enum: [low, medium, high]
                  default: medium
                  example: "high"
      responses:
        '201':
          description: Task berhasil dibuat
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Task created successfully"
                  data:
                    $ref: '#/components/schemas/Task'
        '400':
          description: Validasi input gagal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Token tidak ada atau tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /tasks/{id}:
    get:
      tags:
        - Tasks
      summary: Get single task
      description: Mendapatkan detail satu task berdasarkan ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      responses:
        '200':
          description: Task berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Task retrieved successfully"
                  data:
                    $ref: '#/components/schemas/Task'
        '404':
          description: Task tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Token tidak ada atau tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Tasks
      summary: Update entire task (replace)
      description: Mengganti seluruh data task dengan data baru
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  minLength: 3
                  example: "Updated task title"
                description:
                  type: string
                  example: "Updated description"
                priority:
                  type: string
                  enum: [low, medium, high]
                  example: "low"
                completed:
                  type: boolean
                  example: true
      responses:
        '200':
          description: Task berhasil diupdate
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Task updated successfully"
                  data:
                    $ref: '#/components/schemas/Task'
        '404':
          description: Task tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Validasi input gagal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Token tidak ada atau tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Tasks
      summary: Partially update task
      description: Mengupdate sebagian field task (hanya field yang dikirim)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 3
                  example: "Updated task title"
                description:
                  type: string
                  example: "Updated description"
                priority:
                  type: string
                  enum: [low, medium, high]
                  example: "medium"
                completed:
                  type: boolean
                  example: true
      responses:
        '200':
          description: Task berhasil diupdate
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Task updated successfully"
                  data:
                    $ref: '#/components/schemas/Task'
        '404':
          description: Task tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Validasi input gagal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Token tidak ada atau tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Tasks
      summary: Delete task
      description: Menghapus task berdasarkan ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      responses:
        '204':
          description: Task berhasil dihapus (no content)
        '404':
          description: Task tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Token tidak ada atau tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Mengecek status API server
      responses:
        '200':
          description: Server berjalan normal
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "OK"
                  message:
                    type: string
                    example: "Task Management API is running"
                  timestamp:
                    type: string
                    format: date-time